mcp-weather-ink-suite/client-pi/requirements.txt
mcp>=1.0.0
fastmcp>=0.9.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
Pillow>=10.0.0
httpx>=0.25.0


spidev
gpiozero
RPi.GPIO
